Revert "Revert "Allow shim to handle multiple trusted certificates""
The reasoning given in "Revert "Allow shim to handle multiple trusted
certificates"" is incorrect.  The original commit "Allow shim to
handle multiple trusted certificates" makes no changes to the
behaviour of shim in terms of blacklisted certificate checking.

If it is desired to add support for checking shim_cert and/or
vendor_cert against the certificate blacklists, then this
functionality should be added separately.  It is entirely unrelated to
the ability of shim to handle multiple trusted certificates.

This reverts commit cdbfb5a69ea16142d132517d39c9e96d154c8345.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>