Add proxy loader mechanism
Provide a proxyload() function with the same prototype as LoadImage(),
that can be used to safely create a new image handle for an in-memory
image that can be validated by shim's handle_image().

The new image handle is created by the UEFI platform firmware itself
(by calling LoadImage() on the embedded copy of ProxyLoader.efi).  The
image handle and any installed protocols such as EFI_LOADED_IMAGE are
therefore guaranteed to be valid and safe for use by UEFI platform
functions such as StartImage() and UnloadImage().

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>