ipxe: Allow next loader path to be derived from shim path
Allow loader path to be constructed from the path used to load the
shim itself, e.g.:

    ipxe-shimx64.efi      -> ipxe.efi
    ipxe-shimaa64.efi     -> ipxe.efi
    snponly-shimx64.efi   -> snponly.efi
    snponly-shimaa64.efi  -> snponly.efi

This reduces the complexity of using a signed shim binary to load
iPXE, which (unlike GRUB) has a variety of possible binary names
depending on the requested driver set.  For example, if a site uses
all three of ipxe.efi, intel.efi, and snponly.efi then symlinks can be
used to provide the appropriate shim files:

    # iPXE binaries
    /var/lib/tftpboot/ipxe.efi
    /var/lib/tftpboot/intel.efi
    /var/lib/tftpboot/snponly.efi

    # shim binary (from this repository)
    /var/lib/tftpboot/ipxe-shimx64.efi

    # shim symlinks
    /var/lib/tftpboot/intel-shimx64.efi -> ipxe-shimx64.efi
    /var/lib/tftpboot/snponly-shimx64.efi -> ipxe-shimx64.efi

Unlike the shim binary names, iPXE binary names do not conventionally
include a CPU architecture suffix such as "x64" or "aa64": the
expectation is that if multiple architectures are in use at a site,
then there will be one directory per architecture.  For example:

    /var/lib/tftpboot/x86_64/ipxe.efi
    /var/lib/tftpboot/x86_64/ipxe-shimx64.efi
    /var/lib/tftpboot/arm64/ipxe.efi
    /var/lib/tftpboot/arm64/ipxe-shimaa64.efi

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>